renamed multiple files & reorganized many of the explorer body patterns
