+++ src/app/patterns/body/explorer/ComparisonSummary.svelte
