+++ src/app/patterns/body/explorer/ProportionExplorer.svelte
@@ -4,9 +4,9 @@ import { tweened } from 'svelte/motion';
 import { cubicOut as easing } from 'svelte/easing';
 import { format } from 'd3-format';
 
-import BuildIDComparison from '../elements/BuildIDComparison.svelte';
-import DistributionComparison from '../elements/DistributionComparison.svelte';
-import ComparisonSummary from '../elements/ComparisonSummary.svelte';
+import BuildIDComparison from './BuildIDComparison.svelte';
+import DistributionComparison from './DistributionComparison.svelte';
+import ComparisonSummary from './ComparisonSummary.svelte';
 
 import { formatBuildIDToDateString } from '../utils/formatters';
