+++ src/app/patterns/body/explorer/QuantileExplorer.svelte
@@ -5,10 +5,10 @@ import { tweened } from 'svelte/motion';
 import { cubicOut as easing } from 'svelte/easing';
 import { format } from 'd3-format';
 
-import BuildIDComparison from '../elements/BuildIDComparison.svelte';
-import TotalClientsGraph from '../elements/TotalClientsGraph.svelte';
-import DistributionComparison from '../elements/DistributionComparison.svelte';
-import ComparisonSummary from '../elements/ComparisonSummary.svelte';
+import BuildIDComparison from './BuildIDComparison.svelte';
+// import TotalClientsGraph from './TotalClientsGraph.svelte';
+import DistributionComparison from './DistributionComparison.svelte';
+import ComparisonSummary from './ComparisonSummary.svelte';
 import Violin from '../../../../components/data-graphics/ViolinPlotMultiple.svelte';
 
 import { explorerComparisonSmallMultiple } from '../utils/constants';
