+++ src/app/patterns/errors/DataError.svelte
@@ -73,6 +73,6 @@ export let moreInformation;
     </div>
   {/if}
   <div class="data-error-msg__call-to-action">
-    If you think this is a bug, report this on the <a href='#'>#glam</a> channel on Mozilla's Slack instance or in the <a href='https://github.com/mozilla/glam/issues/new'>GLAM github repository</a>.
+    If you think this is a bug, report this on the <a href='https://mozilla.slack.com/archives/CB1EQ437S'>#glam</a> channel on Mozilla's Slack instance or in the <a href='https://github.com/mozilla/glam/issues/new'>GLAM github repository</a>.
   </div>
 </div>
